feat: drag to expand the table row
