+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/_shared_widget.dart
@@ -8,358 +8,6 @@ import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-/// Only displaying the add row / add column / add column and row button
-///   when hovering on the last row / last column / last cell.
-bool _enableHoveringLogicV2 = true;
-
-class SimpleTableAddRowHoverButton extends StatefulWidget {
-  const SimpleTableAddRowHoverButton({
-    super.key,
-    required this.editorState,
-    required this.tableNode,
-  });
-
-  final EditorState editorState;
-  final Node tableNode;
-
-  @override
-  State<SimpleTableAddRowHoverButton> createState() =>
-      _SimpleTableAddRowHoverButtonState();
-}
-
-class _SimpleTableAddRowHoverButtonState
-    extends State<SimpleTableAddRowHoverButton> {
-  late final interceptorKey =
-      'simple_table_add_row_hover_button_${widget.tableNode.id}';
-
-  SelectionGestureInterceptor? interceptor;
-
-  @override
-  void initState() {
-    super.initState();
-
-    interceptor = SelectionGestureInterceptor(
-      key: interceptorKey,
-      canTap: (details) => !_isTapInBounds(details.globalPosition),
-    );
-    widget.editorState.service.selectionService
-        .registerGestureInterceptor(interceptor!);
-  }
-
-  @override
-  void dispose() {
-    widget.editorState.service.selectionService.unregisterGestureInterceptor(
-      interceptorKey,
-    );
-
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    assert(widget.tableNode.type == SimpleTableBlockKeys.type);
-
-    if (widget.tableNode.type != SimpleTableBlockKeys.type) {
-      return const SizedBox.shrink();
-    }
-
-    final simpleTableContext = context.read<SimpleTableContext>();
-    return ValueListenableBuilder(
-      valueListenable: simpleTableContext.isHoveringOnTableArea,
-      builder: (context, isHoveringOnTableArea, child) {
-        return ValueListenableBuilder(
-          valueListenable: simpleTableContext.hoveringTableCell,
-          builder: (context, hoveringTableCell, child) {
-            bool shouldShow = isHoveringOnTableArea;
-            if (hoveringTableCell != null && _enableHoveringLogicV2) {
-              shouldShow =
-                  hoveringTableCell.rowIndex + 1 == hoveringTableCell.rowLength;
-            }
-            return shouldShow
-                ? Positioned(
-                    bottom: 2 * SimpleTableConstants.addRowButtonPadding,
-                    left: SimpleTableConstants.tableLeftPadding -
-                        SimpleTableConstants.cellBorderWidth,
-                    right: SimpleTableConstants.addRowButtonRightPadding,
-                    child: SimpleTableAddRowButton(
-                      onTap: () => widget.editorState.addRowInTable(
-                        widget.tableNode,
-                      ),
-                    ),
-                  )
-                : const SizedBox.shrink();
-          },
-        );
-      },
-    );
-  }
-
-  bool _isTapInBounds(Offset offset) {
-    final renderBox = context.findRenderObject() as RenderBox?;
-    if (renderBox == null) {
-      return false;
-    }
-
-    final localPosition = renderBox.globalToLocal(offset);
-    final result = renderBox.paintBounds.contains(localPosition);
-
-    return result;
-  }
-}
-
-class SimpleTableAddRowButton extends StatelessWidget {
-  const SimpleTableAddRowButton({
-    super.key,
-    this.onTap,
-  });
-
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return FlowyTooltip(
-      message: LocaleKeys.document_plugins_simpleTable_clickToAddNewRow.tr(),
-      child: GestureDetector(
-        behavior: HitTestBehavior.opaque,
-        onTap: onTap,
-        child: MouseRegion(
-          cursor: SystemMouseCursors.click,
-          child: Container(
-            height: SimpleTableConstants.addRowButtonHeight,
-            margin: const EdgeInsets.symmetric(
-              vertical: SimpleTableConstants.addColumnButtonPadding,
-            ),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(
-                SimpleTableConstants.addRowButtonRadius,
-              ),
-              color: context.simpleTableMoreActionBackgroundColor,
-            ),
-            child: const FlowySvg(
-              FlowySvgs.add_s,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class SimpleTableAddColumnHoverButton extends StatefulWidget {
-  const SimpleTableAddColumnHoverButton({
-    super.key,
-    required this.editorState,
-    required this.node,
-  });
-
-  final EditorState editorState;
-  final Node node;
-
-  @override
-  State<SimpleTableAddColumnHoverButton> createState() =>
-      _SimpleTableAddColumnHoverButtonState();
-}
-
-class _SimpleTableAddColumnHoverButtonState
-    extends State<SimpleTableAddColumnHoverButton> {
-  late final interceptorKey =
-      'simple_table_add_column_hover_button_${widget.node.id}';
-
-  SelectionGestureInterceptor? interceptor;
-
-  @override
-  void initState() {
-    super.initState();
-
-    interceptor = SelectionGestureInterceptor(
-      key: interceptorKey,
-      canTap: (details) => !_isTapInBounds(details.globalPosition),
-    );
-    widget.editorState.service.selectionService
-        .registerGestureInterceptor(interceptor!);
-  }
-
-  @override
-  void dispose() {
-    widget.editorState.service.selectionService.unregisterGestureInterceptor(
-      interceptorKey,
-    );
-
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    assert(widget.node.type == SimpleTableBlockKeys.type);
-
-    if (widget.node.type != SimpleTableBlockKeys.type) {
-      return const SizedBox.shrink();
-    }
-
-    return ValueListenableBuilder(
-      valueListenable: context.read<SimpleTableContext>().isHoveringOnTableArea,
-      builder: (context, isHoveringOnTableArea, _) {
-        return ValueListenableBuilder(
-          valueListenable: context.read<SimpleTableContext>().hoveringTableCell,
-          builder: (context, hoveringTableCell, _) {
-            bool shouldShow = isHoveringOnTableArea;
-            if (hoveringTableCell != null && _enableHoveringLogicV2) {
-              shouldShow = hoveringTableCell.columnIndex + 1 ==
-                  hoveringTableCell.columnLength;
-            }
-            return Positioned(
-              top: SimpleTableConstants.tableTopPadding -
-                  SimpleTableConstants.cellBorderWidth,
-              bottom: SimpleTableConstants.addColumnButtonBottomPadding,
-              right: 0,
-              child: Opacity(
-                opacity: shouldShow ? 1.0 : 0.0,
-                child: SimpleTableAddColumnButton(
-                  onTap: () {
-                    widget.editorState.addColumnInTable(widget.node);
-                  },
-                ),
-              ),
-            );
-          },
-        );
-      },
-    );
-  }
-
-  bool _isTapInBounds(Offset offset) {
-    final renderBox = context.findRenderObject() as RenderBox?;
-    if (renderBox == null) {
-      return false;
-    }
-
-    final localPosition = renderBox.globalToLocal(offset);
-    final result = renderBox.paintBounds.contains(localPosition);
-
-    return result;
-  }
-}
-
-class SimpleTableAddColumnButton extends StatelessWidget {
-  const SimpleTableAddColumnButton({
-    super.key,
-    this.onTap,
-  });
-
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return FlowyTooltip(
-      message: LocaleKeys.document_plugins_simpleTable_clickToAddNewColumn.tr(),
-      child: GestureDetector(
-        behavior: HitTestBehavior.translucent,
-        onTap: onTap,
-        child: MouseRegion(
-          cursor: SystemMouseCursors.click,
-          child: Container(
-            width: SimpleTableConstants.addColumnButtonWidth,
-            margin: const EdgeInsets.symmetric(
-              horizontal: SimpleTableConstants.addColumnButtonPadding,
-            ),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(
-                SimpleTableConstants.addColumnButtonRadius,
-              ),
-              color: context.simpleTableMoreActionBackgroundColor,
-            ),
-            child: const FlowySvg(
-              FlowySvgs.add_s,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class SimpleTableAddColumnAndRowHoverButton extends StatelessWidget {
-  const SimpleTableAddColumnAndRowHoverButton({
-    super.key,
-    required this.editorState,
-    required this.node,
-  });
-
-  final EditorState editorState;
-  final Node node;
-
-  @override
-  Widget build(BuildContext context) {
-    assert(node.type == SimpleTableBlockKeys.type);
-
-    if (node.type != SimpleTableBlockKeys.type) {
-      return const SizedBox.shrink();
-    }
-
-    return ValueListenableBuilder(
-      valueListenable: context.read<SimpleTableContext>().isHoveringOnTableArea,
-      builder: (context, isHoveringOnTableArea, child) {
-        return ValueListenableBuilder(
-          valueListenable: context.read<SimpleTableContext>().hoveringTableCell,
-          builder: (context, hoveringTableCell, child) {
-            bool shouldShow = isHoveringOnTableArea;
-            if (hoveringTableCell != null && _enableHoveringLogicV2) {
-              shouldShow = hoveringTableCell.isLastCellInTable;
-            }
-            return shouldShow
-                ? Positioned(
-                    bottom:
-                        SimpleTableConstants.addColumnAndRowButtonBottomPadding,
-                    right: SimpleTableConstants.addColumnButtonPadding,
-                    child: SimpleTableAddColumnAndRowButton(
-                      onTap: () => editorState.addColumnAndRowInTable(node),
-                    ),
-                  )
-                : const SizedBox.shrink();
-          },
-        );
-      },
-    );
-  }
-}
-
-class SimpleTableAddColumnAndRowButton extends StatelessWidget {
-  const SimpleTableAddColumnAndRowButton({
-    super.key,
-    this.onTap,
-  });
-
-  final VoidCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return FlowyTooltip(
-      message: LocaleKeys.document_plugins_simpleTable_clickToAddNewRowAndColumn
-          .tr(),
-      child: GestureDetector(
-        behavior: HitTestBehavior.translucent,
-        onTap: onTap,
-        child: MouseRegion(
-          cursor: SystemMouseCursors.click,
-          child: Container(
-            width: SimpleTableConstants.addColumnAndRowButtonWidth,
-            height: SimpleTableConstants.addColumnAndRowButtonHeight,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(
-                SimpleTableConstants.addColumnAndRowButtonCornerRadius,
-              ),
-              color: context.simpleTableMoreActionBackgroundColor,
-            ),
-            child: const FlowySvg(
-              FlowySvgs.add_s,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
 class SimpleTableRowDivider extends StatelessWidget {
   const SimpleTableRowDivider({
     super.key,
