+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/simple_table/simple_table_operations/simple_table_delete_operation.dart
@@ -18,8 +18,9 @@ extension TableDeletionOperations on EditorState {
   /// Row 1: |   |   |   |
   Future<void> deleteRowInTable(
     Node node,
-    int index,
-  ) async {
+    int index, {
+    bool inMemoryUpdate = false,
+  }) async {
     assert(node.type == SimpleTableBlockKeys.type);
 
     if (node.type != SimpleTableBlockKeys.type) {
@@ -48,7 +49,12 @@ extension TableDeletionOperations on EditorState {
     if (attributes != null) {
       transaction.updateNode(node, attributes);
     }
-    await apply(transaction);
+    await apply(
+      transaction,
+      options: ApplyOptions(
+        inMemoryUpdate: inMemoryUpdate,
+      ),
+    );
   }
 
   /// Delete a column at the given index.
